ENH: Upgrade visualizer rendering to use q2templates
