+++ q2_emperor/plugin_setup.py
@@ -6,36 +6,12 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
-import qiime
-import skbio
-
-from qiime.plugin import Plugin, Metadata, Str
 
 import q2_emperor
+from ._plot import plot
+
+from qiime.plugin import Plugin, Metadata, Str
 from q2_types import PCoAResults
-from emperor import Emperor
-from os.path import join
-
-
-def plot(output_dir: str, metadata: qiime.Metadata,
-         pcoa: skbio.OrdinationResults, custom_axis: str=None) -> None:
-
-    mf = metadata.to_dataframe()
-
-    output = join(output_dir, 'emperor-required-resources/')
-    viz = Emperor(pcoa, mf, remote='.')
-
-    with open(join(output_dir, 'index.html'), 'w') as f:
-        if custom_axis is not None:
-            # put custom_axis inside a list to workaround the type system not
-            # supporting lists of types
-            html = viz.make_emperor(standalone=True, custom_axes=[custom_axis])
-        else:
-            html = viz.make_emperor(standalone=True)
-        viz.copy_support_files(output_dir)
-        f.write(html)
-
-    return None
 
 
 plugin = Plugin(
